VM Handler, Resource Handler 테스트 프로그램 추가
